--- tests/fail-dep/libc/libc-epoll-blocking.stderr
@@ -1,29 +1,0 @@
-error: Undefined Behavior: Data race detected between (1) non-atomic write on thread `unnamed-ID` and (2) retag read of type `u8` on thread `main` at ALLOC. (2) just happened here
-  --> tests/fail-dep/libc/libc-epoll-blocking.rs:LL:CC
-   |
-LL |         assert_eq!(VAL, 1)
-   |         ^^^^^^^^^^^^^^^^^^ Data race detected between (1) non-atomic write on thread `unnamed-ID` and (2) retag read of type `u8` on thread `main` at ALLOC. (2) just happened here
-   |
-help: and (1) occurred earlier here
-  --> tests/fail-dep/libc/libc-epoll-blocking.rs:LL:CC
-   |
-LL |         unsafe { VAL = 1 };
-   |                  ^^^^^^^
-   = help: retags occur on all (re)borrows and as well as when references are copied or moved
-   = help: retags permit optimizations that insert speculative reads or writes
-   = help: therefore from the perspective of data races, a retag has the same implications as a read or write
-   = help: this indicates a bug in the program: it performed an invalid operation, and caused Undefined Behavior
-   = help: see https://doc.rust-lang.org/nightly/reference/behavior-considered-undefined.html for further information
-   = note: BACKTRACE (of the first span):
-   = note: inside `test_epoll_race` at RUSTLIB/core/src/macros/mod.rs:LL:CC
-note: inside `main`
-  --> tests/fail-dep/libc/libc-epoll-blocking.rs:LL:CC
-   |
-LL |     test_epoll_race();
-   |     ^^^^^^^^^^^^^^^^^
-   = note: this error originates in the macro `assert_eq` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: some details are omitted, run with `MIRIFLAGS=-Zmiri-backtrace=full` for a verbose backtrace
-
-error: aborting due to 1 previous error
-
